Move serialization to LogEvent.getBody(). (#201)
This PR moves the serialization from POJO to JSON into LogEvent#getBody(). This allows the serialization of the event to be offloaded to with the event handler when the HTTP request is built. This should be 100% backwards compatible with any existing EventHandler implementations. Initial benchmarks are much improved over the current master in tight loop simulations.