Rework attribute audience conditions
Addresses concerns of excessive complexity in business logic and logging
strategy for audience conditions on user attributes while maintaing the
same matching behavior. Only a minor changes have been made some of the
log mesages, but should not impact the ability for SDK end-user to
debug.

The `Match` abstraction and `MatchType` class is overly complex and
hindered the ability to add detailed logging in a non-intrusive manner.
The `Match` interface and implementations were replaced by classes that
implement `Condition` interface directly that share common logic.

Additionally, this replaces the `castToValueType` method that performed
unecessary reflection and was difficult to reason about -- the class
that is required for a match operation is known ahead of time.

Further improvements can be made to this code, such as removing
object allocations when condition is evaluated and cleaning up the
associated `ConditionUtils.parseConditions` logic, but that is left
intact to minimize amount of change.