Move JUnit config for ExhaustiveTest to core-api
The `com.optimizely.ab.categories.ExhaustiveTest` JUnit test category
marker is defined in the core-api subproject. It is also only used in the
core-api subproject.

Isolate the configuration to use this class to the core-api subproject
by moving out of the root `build.gradle`. This way, we can remove
core-api's testSources as a dependency of core-httpclient-impl's
testSources, as well as any additional subprojects that are added to
this repository.