Create ProjectConfigManager and refactor for dynamic ProjectConfig updates. (#5)
In preparation for a background ProjectConfig updater it became apparent that initializing Bucketer and DecisionService required a large amount of instance level juggling. Moving to a more stateless implementation of those ProjectConfig consumers made ProjectConfig management much more trackable. The major update as part of that refactor was to update all of the mocked instances.