Implement Event Batching (#312)
* Introduces an EventProcessor interface.
* Introduces a BatchEventProcessor (maybe should name it BatchingEventProcessor)
* Introduces an OptimizelyRule to handle lifecycle management during unit testing.

Buffering events within a queue before dispatching is an optimization that should prevent SDK implementations from exhausting resources while increasing throughput. This implementation relies on a `BlockingQueue` to buffer events received from one-to-many producers. A single consumer thread continuously polls from this queue to build a batch before emitting the batched `LogEvent`.

Note that the `Optimizely` class only knows about the `EventProcessor` and `EventHandler` so that it can gracefully close those implementations, so not strictly necessary. The Optimizely class builds UserEvents via the `UserEventFactory` and passes those events to the `EventProcessor`. The `EventProcessor` emits `LogEvents` to a `NotificationManager<LogEvent>` which is subscribed to by the `EventHandler` for eventual dispatching.